1. 테스트 코드 작성 2. 주민번호 암복호화 관련 로직 수정 3. swagger관련 설정 수정 4. readme 작성 5. retrofit2 적용 (동기, 비동기 구현) 6. properties 값 암호화

+++ src/main/java/com/homework/domain/scrap/Tax.java
@@ -21,7 +21,7 @@ public class Tax {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private Long totalSalary; // 총급여
-    private Long calculatedTax; // 산출세액
+    private Double calculatedTax; // 산출세액
     private Long insuranceAmount; // 보험료
     private Long medicalAmount; // 의료비
     private Long educationAmount; // 교육비
